@@ -1,72 +1,65 @@
 {
-  "name": "web-plugin-demo",
+  "name": "web-plugin",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build-web-component": "./build.sh",
-    "build": "ng build --prod",
+    "build": "ng build",
+    "watch": "ng build --watch --configuration development",
     "test": "ng test",
     "testgithub": "ng test --watch=false --browsers=ChromeHeadless --code-coverage",
-    "buildwebpluginprod": "ng build @tesla-ce/web-plugin --prod",
-    "buildwebpluginteslacore": "ng build web-plugin --prod --output-hashing=none && cat dist/web-plugin/runtime-es5.js dist/web-plugin/polyfills-es5.js dist/web-plugin/main-es5.js > dist/web-plugin/web-plugin.js",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "buildwebpluginprod": "ng build @tesla-ce/web-plugin --configuration production",
+    "buildwebpluginteslacore": "ng build web-plugin --configuration=production --output-hashing=none && cat dist/web-plugin/runtime-es5.js dist/web-plugin/polyfills-es5.js dist/web-plugin/main-es5.js > dist/web-plugin/web-plugin.js",
+    "lint": "ng lint"
   },
-  "private": false,
+  "private": true,
   "dependencies": {
-    "@angular/animations": "~9.1.7",
-    "@angular/cdk": "^9.2.4",
-    "@angular/common": "~9.1.7",
-    "@angular/compiler": "~9.1.7",
-    "@angular/core": "~9.1.7",
-    "@angular/elements": "^9.1.7",
-    "@angular/forms": "~9.1.7",
-    "@angular/localize": "^9.1.7",
-    "@angular/material": "^9.2.3",
-    "@angular/platform-browser": "~9.1.7",
-    "@angular/platform-browser-dynamic": "~9.1.7",
-    "@angular/router": "~9.1.7",
-    "@ngx-pwa/local-storage": "^9.0.3",
-    "@ngx-translate/core": "^12.1.2",
+    "@angular/animations": "~12.1.0",
+    "@angular/cdk": "^12.1.0",
+    "@angular/common": "~12.1.0",
+    "@angular/compiler": "~12.1.0",
+    "@angular/core": "~12.1.0",
+    "@angular/elements": "^12.1.0",
+    "@angular/forms": "~12.1.0",
+    "@angular/material": "^12.1.0",
+    "@angular/platform-browser": "~12.1.0",
+    "@angular/platform-browser-dynamic": "~12.1.0",
+    "@angular/router": "~12.1.0",
+    "@nebular/theme": "8.0.0",
+    "@ngx-pwa/local-storage": "^12.1.0",
+    "@ngx-translate/core": "^12.1.0",
     "@ngx-translate/http-loader": "^4.0.0",
-    "angular-svg-icon": "^9.2.0",
-    "document-register-element": "^1.14.10",
+    "@tesla-ce/icons": "0.0.40",
+    "@tesla-ce/sensors": "0.0.15",
+    "@types/dom-mediacapture-record": "^1.0.11",
+    "angular-svg-icon": "^12.0.0",
     "moment": "^2.29.1",
-    "ngx-moment": "^3.5.0",
-    "rxjs": "~6.5.4",
-    "tslib": "^1.14.1",
-    "zone.js": "~0.10.2"
+    "ngx-moment": "^5.0.0",
+    "rxjs": "~6.6.0",
+    "tslib": "^2.2.0",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "^9.2.0",
-    "@angular-builders/dev-server": "^7.3.1",
-    "@angular-devkit/build-angular": "^0.901.13",
-    "@angular-devkit/build-ng-packagr": "^0.901.13",
-    "@angular/cli": "^9.1.13",
-    "@angular/compiler": "~9.1.7",
-    "@angular/compiler-cli": "^9.1.13",
-    "@biesbjerg/ngx-translate-extract": "^6.1.0",
-    "@types/dom-mediacapture-record": "^1.0.7",
-    "@types/jasmine": "^3.5.14",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^12.19.15",
-    "@webcomponents/webcomponentsjs": "^2.5.0",
-    "codelyzer": "^5.1.2",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^5.0.9",
+    "@angular-devkit/build-angular": "~12.1.4",
+    "@angular-eslint/builder": "12.7.0",
+    "@angular-eslint/eslint-plugin": "12.7.0",
+    "@angular-eslint/eslint-plugin-template": "12.7.0",
+    "@angular-eslint/schematics": "12.7.0",
+    "@angular-eslint/template-parser": "12.7.0",
+    "@angular/cli": "~12.1.4",
+    "@angular/compiler-cli": "~12.1.0",
+    "@types/jasmine": "~3.8.0",
+    "@types/node": "^12.11.1",
+    "@typescript-eslint/eslint-plugin": "4.28.2",
+    "@typescript-eslint/parser": "4.28.2",
+    "eslint": "^7.26.0",
+    "jasmine-core": "~3.8.0",
+    "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~2.1.0",
-    "karma-jasmine": "~3.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
-    "ng-packagr": "^9.1.5",
-    "protractor": "~5.4.3",
-    "stylelint": "^13.9.0",
-    "stylelint-config-standard": "^20.0.0",
-    "ts-node": "~8.3.0",
-    "tslint": "^6.1.3",
-    "typescript": "^3.8.3",
-    "url-loader": "^4.1.1"
+    "karma-coverage": "~2.0.3",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "~1.7.0",
+    "ng-packagr": "^12.1.0",
+    "typescript": "~4.3.2"
   }
 }
