@@ -1,38 +1,22 @@
 {
   "name": "@tesla-ce/web-plugin",
-  "version": "0.0.15",
-  "homepage": "https://www.tesla-ce.eu",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/tesla-ce/angular-webplugin"
-  },
-  "license": "AGPL-3.0-or-later",
-  "dependencies": {
-    "@angular/common": "^9.1.10",
-    "@angular/compiler": "9.1.13",
-    "@angular/compiler-cli": "9.1.13",
-    "@angular/core": "^9.1.10",
-    "@angular/material": "^9.2.3",
-    "@angular/platform-browser": "~9.1.10",
-    "@angular/platform-browser-dynamic": "~9.1.10",
-    "@ngx-pwa/local-storage": "^9.0.3",
-    "@ngx-translate/core": "^12.1.2",
+  "version": "0.0.17",
+  "peerDependencies": {
+    "@angular/common": "^12.1.0",
+    "@angular/core": "^12.1.0",
+    "@angular/material": "^12.1.0",
+    "@ngx-translate/core": "^12.1.0",
     "@ngx-translate/http-loader": "^4.0.0",
-    "@tesla-ce/icons": "0.0.36",
-    "@tesla-ce/sensors": "0.0.11",
-    "angular-svg-icon": "^9.2.0",
-    "document-register-element": "^1.7.2",
-    "moment": "^2.26.0",
-    "ngx-moment": "^3.5.0",
-    "rxjs": "~6.5.4",
-    "tslib": "^1.10.0",
-    "typescript": "^3.8.3"
+    "@ngx-pwa/local-storage": "^12.1.0",
+    "moment": "^2.29.1",
+    "ngx-moment": "^5.0.0",
+    "rxjs": "~6.6.0",
+    "@tesla-ce/icons": "0.0.40",
+    "@tesla-ce/sensors": "0.0.14"
   },
-  "publishConfig": {
-    "access": "public"
-  },
-  "author": "Xavier Baró <xbaro@uoc.edu>",
-  "scripts": {
-    "postinstall": "ngcc"
+  "dependencies": {
+    "tslib": "^2.2.0",
+    "@tesla-ce/icons": "0.0.40",
+    "@tesla-ce/sensors": "0.0.15"
   }
 }
