@@ -1,21 +1,23 @@
+import {Injector, NgModule} from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { NgModule, Injector } from '@angular/core';
-// import { WebPluginModule, WebPluginComponent} from '@tesla-ce/web-plugin';
+
 import { WebPluginModule, WebPluginComponent} from '../../projects/tesla-ce/web-plugin/src/public-api';
 import { createCustomElement } from '@angular/elements';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 @NgModule({
-  declarations: [],
+  declarations: [
+  ],
   imports: [
     BrowserModule,
-    WebPluginModule
+    WebPluginModule,
+    BrowserAnimationsModule
   ],
-  providers: [],
   entryComponents: [WebPluginComponent],
   bootstrap: []
 })
 export class AppModule {
-  constructor(private injector: Injector) { }
+  constructor(private injector: Injector) {}
 
   ngDoBootstrap() {
     const ngElement = createCustomElement(WebPluginComponent, {
@@ -23,5 +25,4 @@ export class AppModule {
     });
     customElements.define('tesla-web-plugin', ngElement);
   }
-
 }
